Moved data_importer to src, tidy/updated other
