+++ src/data_importer.py
@@ -1,11 +1,14 @@
 import pathlib as pl
 import csv
 import sys
+import _csv # for type annotation
+from typing import Union
 from multiprocessing import Pool, cpu_count
 
 from preprocessing.noise_removal import clean_str # type: ignore
 from utils.types import news_info, words_info # type: ignore
-from preprocessing.words_dicts import WordsDicts # type: ignore
+from utils.mappings import out_cols # type: ignore
+from preprocessing.data_handlers import WordsDicts, CsvWriter # type: ignore
 
 def create_clear_buffer(n_procs: int) -> list[list[news_info]]:
     """Create buffer list with n_procs empty lists."""
@@ -14,14 +17,14 @@ def create_clear_buffer(n_procs: int) -> list[list[news_info]]:
         buffer.append([])
     return buffer
 
-def process_batch(articles: list[news_info]) -> list[words_info]:
-    """Clean text and split into list of type/bag of words pairs."""
-    return [(t, clean_str(c).split(" ")) for t, c in articles]
-
-def process_buffer(buffer: list[list[news_info]], n_procs: int) -> list[words_info]:
+def process_buffer(
+    out_obj: Union[CsvWriter, WordsDicts],
+    buffer: list[list[Union[words_info, tuple[str, ...]]]],
+    n_procs: int,
+) -> list[Union[words_info, tuple[str, ...]]]:
     """Multiprocess article buffer, return list of type/bag of words pairs."""
     with Pool(n_procs) as p:
-        data_results = p.map_async(process_batch, buffer)
+        data_results = p.map_async(out_obj.process_batch, buffer)
         data_lists = data_results.get()
         # Concattenate list of lists of words to just list of word_info
         return [article for batch in data_lists for article in batch]
@@ -36,26 +39,21 @@ def set_maxsize():
         except OverflowError:
             maxInt = int(maxInt/10)
 
-def raw_to_words(
-    from_file: pl.Path,
-    to_path: pl.Path,
+def process_lines(
     n_rows: int,
-    incl_name: str,
-    excl_name: str
+    reader: '_csv._reader',
+    out_obj: Union[CsvWriter, WordsDicts],
 ) -> tuple[int, int, int]:
     """Read raw csv file line by line, clean words, count occurrences and dump to json.
     
     Return tuple of n_included, n_excluded, n_skipped
     """
     set_maxsize()
-    
-    out_dicts = WordsDicts(to_path, incl_name, excl_name)
 
     n_procs = cpu_count()
-    batch_sz = max(1, min(10000, n_rows//n_procs))
-    buffer_sz = n_procs * batch_sz
-    # n empty lists
-    buffer = create_clear_buffer(n_procs)
+    batch_sz = max(1, min(10000, n_rows//n_procs)) # At least 1, at most 10000
+    buffer_sz = n_procs * batch_sz # Buffer of n batches of batch_sz
+    buffer = create_clear_buffer(n_procs) # n empty lists
 
     n_read: int = 0 # Count rows that were be read.
     n_skipped: int = 0 # Count skipped rows that could not be read.
@@ -63,50 +61,80 @@ def raw_to_words(
 
     running = True
     i = 0
-    
-    with open(from_file, encoding="utf8") as f:
-        reader = csv.reader(f)
-        next(reader) # skip header
-        row = next(reader)
 
-        while running:
+    while running:
+        try:
+            row = next(reader)
+            # "Progress bar"
+            if i % 5000 == 0:
+                print("Lines read:", i, "...")
+            # Parallel process data if all batches are full
+            if n_read % buffer_sz == 0:
+                out_obj.write(process_buffer(out_obj, buffer, n_procs))
+                buffer = create_clear_buffer(n_procs)
+            # Read and save line
             try:
-                row = next(reader)
-                if i % 5000 == 0:
-                    print("Lines read:", i, "...") # "progress bar"
-
-                # Parallel process data if all batches are full
-                if n_read % buffer_sz == 0:
-                    out_dicts.add_words(process_buffer(buffer, n_procs))
-                    buffer = create_clear_buffer(n_procs)
-
-                # Read and save line
-                try:
-                    type_ = row[3]
-                    content = row[5]
-                    # Add article to appropriate batch
+                # Add article to appropriate batch
+                extracted = out_obj.extract(row)
+                if extracted != ():
                     buffer_index = (n_read % buffer_sz)//batch_sz
                     batch = buffer[buffer_index]
-                    batch.append((type_, content))
-                    n_read += 1
-                # Or skip row if either type or content cannot be read
-                except:
-                    n_skipped += 1 # ERROR HERE TODO
-                    print("batch add fail", i, n_skipped)
-                
-                # Break when target rows reached
-                if i >= n_rows:
-                    running = False
-                i += 1
-
-            except: # No row to read (or other error)
+                    batch.append(extracted)
+                n_read += 1
+            # Or skip row if either type or content cannot be read
+            except:
+                n_skipped += 1
+            # Break when target rows reached
+            if i >= n_rows:
                 running = False
-        
-        # Flush what remains in the buffer
-        out_dicts.add_words(process_buffer(buffer, n_procs))
-
+            i += 1
+        except: # No row to read (or other error)
+            running = False
+    # Flush what remains in the buffer
+    out_obj.write(process_buffer(out_obj, buffer, n_procs))
     # Export as json
-    out_dicts.stem()
-    out_dicts.export_json()
+    out_obj.finalize()
+    return out_obj.n_incl, out_obj.n_excl, n_skipped
 
-    return out_dicts.n_incl, out_dicts.n_excl, n_skipped
+def raw_to_words(
+    from_file: pl.Path,
+    to_path: pl.Path,
+    n_rows: int,
+    incl_name: str,
+    excl_name: str
+) -> None:
+    """Read raw csv file line by line, clean words, count occurrences and dump to json.
+    
+    Return tuple of n_included, n_excluded, n_skipped.
+    """
+    to_path.mkdir(parents=True, exist_ok=True) # Create dest folder if it does not exist
+    out_dicts = WordsDicts(to_path, incl_name, excl_name)
+    with open(from_file, encoding="utf8") as ff:
+        reader = csv.reader(ff)
+        next(reader) # skip header
+        n_incl, n_excl, n_skipped = process_lines(n_rows, reader, out_obj=out_dicts)
+    print(f"{n_incl + n_excl} rows read, \n {n_incl} were included \n {n_excl} were excluded \n {n_skipped} were skipped \n JSON files were written to {to_path}") # PACK away in classes maybe TODO
+    if n_incl + n_excl != n_rows:
+        raise ValueError("Count numbers do not match")
+
+def reduce_raw(
+    from_file: pl.Path,
+    to_path: pl.Path,
+    n_rows: int
+) -> None:
+    """Stream-read corpus, process and stream-write to new file.
+    
+    Return tuple of n_included, n_excluded, n_skipped.
+    """
+    to_path.mkdir(parents=True, exist_ok=True) # Create dest folder if it does not exist
+    with open(from_file, encoding="utf8") as ff:
+        reader = csv.reader(ff)
+        row = next(reader) # Get headers
+        with open(to_path / "reduced_corpus.csv", 'w', encoding="utf8") as tf:
+            csv_writer = csv.writer(tf)
+            csv_writer.writerow(out_cols) # Write headers
+            writer = CsvWriter(csv_writer)
+            n_incl, n_excl, n_skipped = process_lines(n_rows, reader, out_obj=writer)
+    print(f"{n_incl + n_excl} rows read, \n {n_incl} were included \n {n_excl} were excluded \n {n_skipped} were skipped \n Reduced csv data file was written to {to_path}")
+    if n_incl + n_excl != n_rows:
+        raise ValueError("Count numbers do not match")
